Exercise complete

+++ app/Http/Controllers/Admin/PostController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\Request;
 
 use App\Post;
 use App\Category;
+use App\Tag;
 
 
 class PostController extends Controller
@@ -21,7 +22,6 @@ class PostController extends Controller
     {
         $posts= Post::all();
 
-
         return view('admin.posts.index', compact('posts'));
 
     }
@@ -34,8 +34,11 @@ class PostController extends Controller
     public function create()
     {
         $categories = Category::all();
+        $tags= Tag::all();
 
-        return view('admin.posts.create', compact('categories'));
+        return view('admin.posts.create', compact('categories', 'tags'));
+
+
     }
 
     /**
@@ -50,7 +53,8 @@ class PostController extends Controller
             [
             'title' => 'required|max:255',
             'content' => 'required|min:8|max:100',
-            'category_id' => 'required|exists:categories,id'
+            'category_id' => 'required|exists:categories,id',
+            'tags' => 'exists:tags,id',
             ],
             // L'array sottostante equivale ad un messaggio di errore personalizzato,
             // Lo si può utilizzare per cambiare il soggetto dell'errore es 'name.required' => 'The name field is required.'
@@ -58,12 +62,15 @@ class PostController extends Controller
                 'title.required' => 'LoL, you forgot the title.',
                 'content.min' => "C'mon man, you're almost there!",
                 'content.required'=> 'LoL, you also forgot the content.',
-                'category_id.exists' => 'La categoria selected don\'t exists'
+                'category_id.exists' => 'La categoria selected don\'t exists',
+                'tags' => 'The tag selected don\'t exists'
             ]
         );
             $postData = $request->all();
             $newPost = new Post();
+
             $newPost->fill($postData);
+
             $slug = Str::slug($newPost->title);
             $alternativeSlug = $slug;
             $postFound = Post::where('slug', $alternativeSlug)->first();
@@ -74,6 +81,14 @@ class PostController extends Controller
                 $postFound = Post::where('slug', $alternativeSlug)->first();
             }
             $newPost->slug = $alternativeSlug;
+            $newPost->save();
+
+            // add tags
+            if(array_key_exists('tags', $postData)){
+                $newPost->tags()->sync($postData['tags']);
+            }
+
+
             $newPost->save();
             return redirect()->route('admin.posts.index');
 
@@ -94,8 +109,9 @@ class PostController extends Controller
         }
 
         $category = Category::find($post->category_id);
+        $tags= Tag::find($post);
 
-        return view('admin.posts.show', compact('post', 'category'));
+        return view('admin.posts.show', compact('post', 'category', 'tags'));
 
     }
 
@@ -105,13 +121,14 @@ class PostController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(Post $post)
     {
-        $post=Post::find($id);
+        // $post=Post::find($id);
 
         $categories = Category::all();
+        $tags= Tag::all();
 
-        return view('admin.posts.edit', compact('post', 'categories'));
+        return view('admin.posts.edit', compact('post', 'categories', 'tags'));
     }
 
     /**
@@ -126,7 +143,8 @@ class PostController extends Controller
         $request->validate([
             'title' => 'required|max:255',
             'content' => 'required|min:8',
-            'category_id' => 'required|exists:categories,id'
+            'category_id' => 'required|exists:categories,id',
+            'tags' => 'exists:tags,id'
         ],
         // L'array sottostante equivale ad un messaggio di errore personalizzato,
         // Lo si può utilizzare per cambiare il soggetto dell'errore es 'name.required' => 'The name field is required.'
@@ -134,7 +152,8 @@ class PostController extends Controller
             'title.required' => 'LoL, you forgot the title.',
             'content.min' => "C'mon man, you're almost there!",
             'content.required'=> 'LoL, you also forgot the content.',
-            'category_id.exists' => 'La categoria selected don\'t exists'
+            'category_id.exists' => 'La categoria selected don\'t exists',
+            'tags' => 'The tag selected don\'t exists'
         ]
     );
             $postData = $request->all();
@@ -150,6 +169,14 @@ class PostController extends Controller
                 $postFound = Post::where('slug', $alternativeSlug)->first();
             }
             $post->slug = $alternativeSlug;
+
+            // add tags
+            if(array_key_exists('tags', $postData)){
+                $post->tags()->sync($postData['tags']);
+            } else {
+                $post->tags()->sync([]);
+            }
+
             $post->update();
             return redirect()->route('admin.posts.index');
     }
@@ -160,10 +187,13 @@ class PostController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Post $post)
     {
-        $post=Post::find($id);
-        $post->delete();
-        return redirect()->route('admin.posts.index', compact('post')) ;
+        if($post){
+            $post->tags()->sync([]);
+            $post->delete();
+        }
+
+        return redirect()->route('admin.posts.index') ;
     }
 }
